requirements.txt
carla==0.9.13
deap==1.4.1
docker==5.0.3
keras==2.6.0
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==4.7.0.72
pandas==1.1.5
Pillow==8.4.0
pygame==2.3.0
scikit-learn==0.24.2
scipy==1.5.4
Shapely==1.8.5.post1

torch~=1.10.1
transformers~=4.18.0
aiofiles~=0.8.0